+++ src/meta.ts
@@ -3,6 +3,6 @@
 
 export { META_KEY, META_VERSION, starterVariants, defaultMeta, loadMeta, persistMeta, updateMeta, resetMeta, deriveBredEntry } from './meta/store';
 export { buySeed, buySeedAndGerminate, buyPoolItem, germinateSeed, germinateVariant, enqueueCross, advanceCrossMaturation, consumeSeed, consumeSeedAndEnqueueCross, addNektar, isCrossReady, buySeedling, plantSeedlingIntoPot, buyRearingSlot } from './meta/economy';
-export { reserveRunId, beginRun, applyRunEnd, recordRunEnd, registerVariant, keepCross, toggleLoadout, canonicalVariantId, payRunResume } from './meta/run';
+export { reserveRunId, beginRun, applyRunEnd, recordRunEnd, registerVariant, keepCross, toggleLoadout, canonicalVariantId, payRunResume, deriveRunStats } from './meta/run';
 export { enqueueBrood, claimBrood, readyBroods } from './meta/run';
 export { deriveLoanPlant, isLoanVariant, LOAN_PLANT_ID } from './meta/loan';